+++ sentry-opentelemetry/sentry-opentelemetry-bootstrap/src/main/java/io/sentry/opentelemetry/SentryWeakSpanStorage.java
@@ -1,9 +1,11 @@
 package io.sentry.opentelemetry;
 
+import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.context.internal.shaded.WeakConcurrentMap;
 import io.sentry.ISentryLifecycleToken;
 import io.sentry.util.AutoClosableReentrantLock;
+import java.lang.ref.WeakReference;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -34,6 +36,8 @@ public final class SentryWeakSpanStorage {
   // weak keys, spawns a thread to clean up values that have been garbage collected
   private final @NotNull WeakConcurrentMap<SpanContext, IOtelSpanWrapper> sentrySpans =
       new WeakConcurrentMap<>(true);
+  private volatile @NotNull WeakReference<IOtelSpanWrapper> lastKnownRootSpan =
+      new WeakReference<>(null);
 
   private SentryWeakSpanStorage() {}
 
@@ -43,11 +47,45 @@ public final class SentryWeakSpanStorage {
 
   public void storeSentrySpan(
       final @NotNull SpanContext otelSpan, final @NotNull IOtelSpanWrapper sentrySpan) {
+    System.out.println("storing span: " + sentrySpan.getOperation());
     this.sentrySpans.put(otelSpan, sentrySpan);
+    if (shouldStoreSpanAsRootSpan(sentrySpan)) {
+      System.out.println("storing span as last known root: " + sentrySpan.getOperation());
+      lastKnownRootSpan = new WeakReference<>(sentrySpan);
+    }
+  }
+
+  private boolean shouldStoreSpanAsRootSpan(final @NotNull IOtelSpanWrapper sentrySpan) {
+    if (!sentrySpan.isRoot()) {
+      return false;
+    }
+
+    final @Nullable IOtelSpanWrapper previousRootSpan = getLastKnownUnfinishedRootSpan();
+    if (previousRootSpan == null) {
+      return true;
+    }
+
+    final @Nullable Attributes attributes = previousRootSpan.getOpenTelemetrySpanAttributes();
+    if (attributes == null) {
+      return true;
+    }
+
+    final @Nullable Boolean isCreatedViaSentryApi =
+        attributes.get(InternalSemanticAttributes.CREATED_VIA_SENTRY_API);
+    return isCreatedViaSentryApi != null && isCreatedViaSentryApi == true;
+  }
+
+  public @Nullable IOtelSpanWrapper getLastKnownUnfinishedRootSpan() {
+    final @Nullable IOtelSpanWrapper span = lastKnownRootSpan.get();
+    if (span != null && !span.isFinished()) {
+      return span;
+    }
+    return null;
   }
 
   @TestOnly
   public void clear() {
     sentrySpans.clear();
+    lastKnownRootSpan.clear();
   }
 }